+++ lib/middleware/index.ts
@@ -1,13 +1,16 @@
 export * from './Auth';
 export {
     CacheControl,
+    type TriFrostCacheControlOptions,
     Sym_TriFrostMiddlewareCacheControl
 } from './CacheControl';
 export {
     Cors,
+    type TriFrostCorsOptions,
     Sym_TriFrostMiddlewareCors
 } from './Cors';
 export {
     Security,
+    type TriFrostSecurityOptions,
     Sym_TriFrostMiddlewareSecurity
 } from './Security';